requirements.txt
os
PIL 
pandas
geopandas
geopy
matplotlib.pyplot
plotly_express 
tqdm
